Allow onValidateContact listener to give feedback
If a plugin listening to the `onValidateContact` event wants to stop contact form submission, it is supposed to return an exception. Currently, after submission is halted, user is sent back to a blank contact form page, stripped of all form elements with no error message indicating why the submission was unsuccessful.

This change resolves the problem by copying the actions for submissions failing native form validation, which is to redirect to contact form pre-filled with submitted values and displaying exception message, hopefully a descriptive explanation of why submission failed.

Relates to Joomla! 4 pull request: https://github.com/joomla/joomla-cms/pull/37540/ This is for anyone looking for a manual fix for Joomla! 3, as this is not a bug fix and will not be merged into Joomla! 3.